move experiment classes around
