+++ all/experiments/single_env_experiment_test.py
@@ -6,6 +6,7 @@ from all.environments import GymEnvironment
 from all.experiments import SingleEnvExperiment
 from all.logging import Writer
 
+
 # pylint: disable=protected-access
 class MockWriter(Writer):
     def __init__(self, experiment, label, write_loss):
